// platforms/web/client/core/src/sampler/snapshot/protocol-gen/org/tiqian/protocol/TableInput.ts
// Generated by the reflaxe TypeScript target. Do not edit.

import { TableMetricRow } from "./TableMetricRow.js";
import { TableProbe } from "./TableProbe.js";

export class TableInput {
  public replayStrings: string[];
  public metrics: TableMetricRow[];
  public probes: TableProbe[];
  public faces: string[];
  public typographies: string[];
  public valueStyles: string[];
  public fontPreloads: string[];
  public revisionsText: string;

  constructor(replayStrings: string[], metrics: TableMetricRow[], probes: TableProbe[], faces: string[], typographies: string[], valueStyles: string[], fontPreloads: string[], revisionsText: string) {
    this.replayStrings = replayStrings;
    this.metrics = metrics;
    this.probes = probes;
    this.faces = faces;
    this.typographies = typographies;
    this.valueStyles = valueStyles;
    this.fontPreloads = fontPreloads;
    this.revisionsText = revisionsText;
  }
}