// platforms/web/client/core/src/sampler/snapshot/protocol-gen/org/tiqian/protocol/TableMetricRow.ts
// Generated by the reflaxe TypeScript target. Do not edit.

export class TableMetricRow {
  public serializedFamilies: string;
  public fontWeight: number;
  public italic: boolean;
  public role: string;
  public faceSelectionText: string;
  public valuesEm: (number | null)[];

  constructor(serializedFamilies: string, fontWeight: number, italic: boolean, role: string, faceSelectionText: string, valuesEm: (number | null)[]) {
    this.serializedFamilies = serializedFamilies;
    this.fontWeight = fontWeight;
    this.italic = italic;
    this.role = role;
    this.faceSelectionText = faceSelectionText;
    this.valuesEm = valuesEm;
  }
}